backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
anthropic>=0.40.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
